pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["RViktorV <rgavih@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
black = "^24.8.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
phonenumbers = "^8.13.45"
django-countries = "^7.6.1"
django-phonenumber-field = "^8.0.0"
django = "4.2"
rest-framework-simplejwt = "^0.0.2"
load-dotenv = "^0.1.0"
djangorestframework-simplejwt = "^5.3.1"
pytest = "^8.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
